+++ UI.java
@@ -88,7 +88,7 @@ public class UI extends UIConstants{
       String firstName = getString("Criminal first name");
       String lastName = getString("Criminal Last name");
       Date dateOfBirth = makeDate("Criminal Date of Birth");
-      int age = getInt("Criminal's age");
+      long age = getLong("Criminal's age");
       String race = getString("Criminal Race");
       double weight = getDouble("Criminal Weight");
       double height = getDouble("Criminal Height");
